feat: profile-edit api/model 세그먼트 추가 및 ui 분리
