+++ src/features/profile-edit/index.ts
@@ -1,1 +1,2 @@
+export { useUpdateProfile } from './model/profile-edit.mutations';
 export { EditProfileModal } from './ui/edit-profile-modal/edit-profile-modal';
